Correction to make the code compile with MSVC.
It seems to make a bit more sense now (as previously
the rounddown seemed to be left abbandoned) but keep in mind
that this correction is a result of random guess and NOT
the result of me understanding what is happening in this code.

This code should be double checked if it does what it was intended to do
and if it produces the right result.